Got rid of effective max, as it offers no advantage. Improved band on levlimopt, compared to postgres. Made max an int.

+++ src/validate_max.h
@@ -4,7 +4,7 @@ This code is common to algorithms with limits.
 */
 
     if (args->args[2]) {
-        long long user_max = *(reinterpret_cast<long long *>(args->args[2]));
+        int user_max = static_cast<int>(*(reinterpret_cast<long long *>(args->args[2])));
         if (user_max < 0) {
             set_error(error, "Maximum edit distance cannot be negative.");
             // We can either return 0 or return null if given a bad max edit distance. This policy is configurable
